Split test_features, copy task related tests to another test script
